Modify office logic to match on remote + other
