+++ index.yaml
@@ -47,6 +47,14 @@ indexes:
 - kind: LoveCount
   properties:
   - name: department
+  - name: office
+  - name: week_start
+  - name: sent_count
+    direction: desc
+
+- kind: LoveCount
+  properties:
+  - name: department
   - name: week_start
   - name: sent_count
     direction: desc
