+++ offices.yaml
@@ -1,3 +1,3 @@
-- Office 1
-- Office 2
-- Office 3
+- SF Office
+- Hamburg Office
+- London Office
